fix(js): move confidential transfer helpers to /confidential subpath
Importing @solana-program/token-2022 eagerly resolved @solana/zk-sdk via
the root barrel, throwing ERR_MODULE_NOT_FOUND even when confidential
transfer helpers were never used.

Move the zk-sdk-dependent helpers to a dedicated /confidential subpath so
the root entry is free of @solana/zk-sdk. Confidential transfer consumers
import from @solana-program/token-2022/confidential and install the
optional peer dependency.
