+++ clients/js/src/confidential.ts
@@ -1,0 +1,2 @@
+export * from './confidentialTransferHelpers';
+export * from './confidentialTransferKeys';
