+++ clients/js/src/index.ts
@@ -4,8 +4,6 @@ export * from './generated';
 export { type BatchInstruction, getBatchInstruction, parseBatchInstruction } from './batch';
 
 export * from './amountToUiAmount';
-export * from './confidentialTransferHelpers';
-export * from './confidentialTransferKeys';
 export * from './getInitializeInstructionsForExtensions';
 export * from './getTokenSize';
 export * from './getMintSize';
