+++ clients/js/test/extensions/confidentialTransfer/confidentialTransferHelpers.test.ts
@@ -1,7 +1,7 @@
 import { generateKeyPairSigner } from '@solana/kit';
 import { AeKey, ElGamalKeypair } from '@solana/zk-sdk/bundler';
 import test from 'ava';
-import { getCreateConfidentialTransferAccountInstructionPlan } from '../../../src';
+import { getCreateConfidentialTransferAccountInstructionPlan } from '../../../src/confidential';
 import { createDefaultSolanaClient } from '../../_setup';
 
 test('it rejects create helper authority that does not match the owner ATA flow', async t => {
