+++ clients/js/test/extensions/confidentialTransfer/getConfidentialTransferInstructionPlan.test.ts
@@ -1,5 +1,6 @@
 import test from 'ava';
-import { fetchToken, getConfidentialTransferInstructionPlan } from '../../../src';
+import { fetchToken } from '../../../src';
+import { getConfidentialTransferInstructionPlan } from '../../../src/confidential';
 import {
     createConfidentialMint,
     createConfidentialTokenAccount,
