+++ clients/js/typedoc.json
@@ -1,5 +1,5 @@
 {
-    "entryPoints": ["src/index.ts"],
+    "entryPoints": ["src/index.ts", "src/confidential.ts"],
     "includeVersion": true,
     "readme": "none",
     "out": "docs"
